+++ src/components/graph/widgets/TextPreviewWidget.vue
